finalizando

+++ src/main/java/com/vsoares/infosw/service/PlanetServiceImpl.java
@@ -2,6 +2,7 @@ package com.vsoares.infosw.service;
 
 import com.vsoares.infosw.model.Planet;
 import com.vsoares.infosw.repository.PlanetRepository;
+import com.vsoares.infosw.utils.StringUtils;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,17 +20,23 @@ public class PlanetServiceImpl implements PlanetService {
     private SwapiService swapiService;
 
     @Override
-    public Planet addPlanet(Planet planet) {
-        if(planet.getName() == null){
+    public Planet addPlanet(Planet newPlanet) {
+        if (newPlanet.getName() == null) {
             throw new NullPointerException("name");
         }
-        planet.setName(planet.getName()
-                        .replaceAll("[^a-zA-Z0-9\\s+]", "")
-                        .toUpperCase()
-                        .stripLeading()
-                        .stripTrailing());
-//        planet.setId(planet.getName().hashCode());
-        long appearances = swapiService.getAppearances(planet.getName());
+
+        String name = StringUtils.normalizeName(newPlanet.getName());
+
+        Planet planet = planetRepository.findByName(name);
+        if (planet != null) {
+            planet.setClimate(newPlanet.getClimate());
+            planet.setTerrain(newPlanet.getTerrain());
+        } else {
+            planet = newPlanet;
+            planet.setName(name);
+            planet.createId();
+        }
+        long appearances = swapiService.getAppearances(name);
         planet.setAppearances(appearances);
         planetRepository.save(planet);
         return planet;
@@ -37,33 +44,35 @@ public class PlanetServiceImpl implements PlanetService {
 
     @Override
     public List<Planet> listPlanets() {
-        List<Planet> planetas = planetRepository.findAll();
-        return  planetRepository.findAll();
+        return planetRepository.findAll();
     }
 
     @Override
     public Planet getByName(String name) {
+        name = StringUtils.normalizeName(name);
         return planetRepository.findByName(name);
     }
 
     @Override
     public Planet getById(String id) {
-        Optional<Planet> planet = planetRepository.findById(new ObjectId(id));
-        if(planet.isPresent()){
-            return planet.get();
-        } else {
-            return null;
+        if (ObjectId.isValid(id)) {
+            Optional<Planet> planet = planetRepository.findById(new ObjectId(id));
+            if (planet.isPresent()) {
+                return planet.get();
+            }
         }
+        return null;
     }
 
     @Override
     public Boolean removePlanet(String id) {
-        Optional<Planet> planet = planetRepository.findById(new ObjectId(id));
-        if(planet.isPresent()){
-            planetRepository.delete(planet.get());
-            return true;
-        } else {
-            return false;
+        if (ObjectId.isValid(id)) {
+            Optional<Planet> planet = planetRepository.findById(new ObjectId(id));
+            if (planet.isPresent()) {
+                planetRepository.delete(planet.get());
+                return true;
+            }
         }
+        return false;
     }
 }
